# API-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Freelance Marketplace.",
        "title": "GigForge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "gigforge.jasim.online",
    "basePath": "/",
    "paths": {
...
        "/client/contracts": {
            "get": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Get all contracts for a specific client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Get all contracts for a client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved contracts",
                        "schema": {
                            "$ref": "#/definitions/Job.GetAllContractsForClientRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/client/contracts/invoices/{contractID}": {
            "get": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Get all invoices for a specific contract",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Get all invoices of a contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "contractID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved invoices",
                        "schema": {
                            "$ref": "#/definitions/Job.GetInvoiceContractRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/client/contracts/{id}": {
            "get": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Get details of a specific contract",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Get a contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved contract",
                        "schema": {
                            "$ref": "#/definitions/Job.GetOneContractForClientRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/client/job": {
            "get": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Get jobs posted by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get my jobs",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved jobs",
                        "schema": {
                            "$ref": "#/definitions/Job.GetMyJobsRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Edit an existing job listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Edit a job",
                "parameters": [
                    {
                        "description": "Job details",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.PostJob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully edited job",
                        "schema": {
                            "$ref": "#/definitions/Job.PostjobRes"
                        }
                    },
                    "400": {
                        "description": "Error validating request body",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Create a new job listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Post a job",
                "parameters": [
                    {
                        "description": "Job details",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.PostJob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully posted job",
                        "schema": {
                            "$ref": "#/definitions/Job.PostjobRes"
                        }
                    },
                    "400": {
                        "description": "Error validating request body",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/client/job/proposals/{id}": {
            "get": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Get all proposals for a specific job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get proposals of a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved proposals",
                        "schema": {
                            "$ref": "#/definitions/Job.GJPRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/client/job/send-offer": {
            "post": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Send a job offer to a freelancer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Send a job offer",
                "parameters": [
                    {
                        "description": "Offer details",
                        "name": "offer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.SendOffer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully sent offer",
                        "schema": {
                            "$ref": "#/definitions/Job.SendOfferRes"
                        }
                    },
                    "400": {
                        "description": "Error validating request body",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/jobs/payment": {
            "get": {
                "description": "Get details of a payment contract",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get payment contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "invoiceID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment contract details",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Job.GetAllContractsForClientRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "contracts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Job.Contracts"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "Job.GetInvoiceContractRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Status": {
                    "type": "integer"
                },
                "invoices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Job.Invoices"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClientAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "FreelancerAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}